Add files via upload
The idea is to develop a program that will aid users to create organized grocery lists. In order to tackle the problem of disorganized grocery lists, we will be predominantly using the dictionary data structure. In this manner, we can use key-value (food category-food item(s)) pairs to assign different grocery items to specific categories. For example, regarding the key name/food category name “vegetables”, the user can input corresponding food items, such as “spinach” and “lettuce”. Furthermore, if there is more than one item per category that a user would like to input, he/she will be able to do so by inputting a list of food items that will be assigned to the corresponding category. On the subject of user input, we will of course be incorporating the input function in order to allow users to freely add and remove food items from categories. The inputting of food items will be done through strings typed by the user; in other words, food items such as spinach or lettuce will be typed as a sequence of characters.